requirements.txt
click==8.1.3
numpy==1.23.3
Pillow==9.2.0
tqdm==4.64.1
